+++ notify-node/src/main/java/com/dwarfeng/notify/node/configuration/ExceptionCodeOffsetConfiguration.java
@@ -13,16 +13,27 @@ public class ExceptionCodeOffsetConfiguration {
     private int exceptionCodeOffset;
     @Value("${notify.exception_code_offset.subgrade}")
     private int subgradeExceptionCodeOffset;
-    @Value("${notify.exception_code_offset.snowflake}")
-    private int snowflakeExceptionCodeOffset;
+    @Value("${notify.exception_code_offset.spring_telqos}")
+    private int springTelqosExceptionCodeOffset;
+    @Value("${notify.exception_code_offset.spring_terminator}")
+    private int springTerminatorExceptionCodeOffset;
     @Value("${notify.exception_code_offset.dwarfeng_datamark}")
     private int dwarfengDatamarkExceptionCodeOffset;
 
     @PostConstruct
     public void init() {
         ServiceExceptionCodes.setExceptionCodeOffset(exceptionCodeOffset);
-        com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.setExceptionCodeOffset(subgradeExceptionCodeOffset);
-        com.dwarfeng.sfds.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(snowflakeExceptionCodeOffset);
-        com.dwarfeng.datamark.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(dwarfengDatamarkExceptionCodeOffset);
+        com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.setExceptionCodeOffset(
+                subgradeExceptionCodeOffset
+        );
+        com.dwarfeng.springtelqos.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                springTelqosExceptionCodeOffset
+        );
+        com.dwarfeng.springterminator.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                springTerminatorExceptionCodeOffset
+        );
+        com.dwarfeng.datamark.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                dwarfengDatamarkExceptionCodeOffset
+        );
     }
 }